Separate internal_name from title
Taxons have an internal name and an external name, both of which are useful.

I originally thought we could have a general purpose name and decide at parse
time, but I think it's better to expose both of them.

The previous code didn't work for internal_name because it lives in the details
hash.