Document project archiving
We are archiving this gem because the functionality has now been ported
to the content-tagger project. This is a change to the readme that will remain in
posterity once the repository is archived.

This code was ported to content-tagger as that was the only application
that was using this gem.